% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_plot}
\alias{add_plot}
\title{Add a ggplot2 plot to an SVG document, replacing a target element}
\usage{
add_plot(target, doc, plot_file, dpi = 150)
}
\arguments{
\item{target}{An xml2 node corresponding to the SVG element to be replaced}

\item{doc}{An xml2 SVG document}

\item{plot_file}{Path to the SVG file of the ggplot2 plot}

\item{dpi}{The resolution to use when interpreting pixel units}
}
\value{
The modified xml2 SVG document (doc) with the plot added
}
\description{
Add a ggplot2 plot to an SVG document, replacing a target element
}
