% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_paths}
\alias{check_paths}
\title{Check if file paths exist}
\usage{
check_paths(paths)
}
\arguments{
\item{paths}{A character vector of file paths to check.}
}
\value{
No return value. Throws an error if any file does not exist.
}
\description{
Check if file paths exist
}
