% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw}
\alias{draw}
\title{Replace an svg object by a ggplot2 graph}
\usage{
draw(input_svg, output_svg, plots, dpi = 150)
}
\arguments{
\item{input_svg}{Path to the input svg file}

\item{output_svg}{Path to the output svg file}

\item{plots}{A named list of ggplot2 objects. The names should correspond to
the ids of the svg elements to be replaced.}

\item{dpi}{The resolution to use when rendering the ggplot2 objects.}
}
\description{
Replace an svg object by a ggplot2 graph
}
