% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_element}
\alias{find_element}
\title{Find an element by id in an SVG document}
\usage{
find_element(doc, id)
}
\arguments{
\item{doc}{An xml2 SVG document}

\item{id}{The id of the element to find}
}
\value{
The xml2 node corresponding to the element with the given id
}
\description{
Find an element by id in an SVG document
}
