% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_doc_unit}
\alias{get_doc_unit}
\title{Get the unit used in the SVG document}
\usage{
get_doc_unit(doc)
}
\arguments{
\item{doc}{An xml2 SVG document}
}
\value{
The unit used in the SVG document (e.g., "px", "mm", "cm", "in")
}
\description{
Get the unit used in the SVG document
}
