% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{insert_image}
\alias{insert_image}
\title{Insert a raster image (PNG/JPG) into an SVG document, replacing a target element}
\usage{
insert_image(doc, label, image_file, dpi = 150)
}
\arguments{
\item{doc}{An SVG document}

\item{label}{The label of the target element to be replaced}

\item{image_file}{Path to the PNG or JPG image to be inserted}

\item{dpi}{The resolution to use when interpreting pixel units (in the template svg)}
}
\value{
The modified SVG document (doc) with the image added and the target removed
}
\description{
Insert a raster image (PNG/JPG) into an SVG document, replacing a target element
}
