% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{insert_svg}
\alias{insert_svg}
\title{Add an element to an SVG document, replacing a target element}
\usage{
insert_svg(doc, label, insert_file, dpi = 150)
}
\arguments{
\item{doc}{An xml2 SVG document}

\item{label}{The label of the target element to be replaced}

\item{dpi}{The resolution to use when interpreting pixel units}

\item{svg_file}{Path to the SVG file to be inserted}
}
\value{
The modified xml2 SVG document (doc) with the svg file added and the target removed
}
\description{
Add an element to an SVG document, replacing a target element
}
