% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{insert_text}
\alias{insert_text}
\title{Replace "{}" in a text element by provided values}
\usage{
insert_text(doc, label, values)
}
\arguments{
\item{doc}{An SVG document}

\item{label}{The label of the text element to edit}

\item{values}{A character vector to replace each "{}" in order}
}
\value{
The modified SVG document (doc) with the text inserted
}
\description{
Replace "{}" in a text element by provided values
}
