% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgedit-package.R
\docType{package}
\name{svgedit-package}
\alias{svgedit}
\alias{svgedit-package}
\title{svgedit: Edit SVG files in R to insert ggplot2 plots and text}
\description{
This package provides functions to edit SVG files created in Inkscape by replacing placeholders (e.g. a rectangle element) by ggplot2 plots or images, and editing text boxes. This automates the creation of figures with complex layouts from R code.
}
\author{
\strong{Maintainer}: Daniel Thedie \email{daniel.thedie@ed.ac.uk} (\href{https://orcid.org/0000-0002-1352-7245}{ORCID})

}
\keyword{internal}
