% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{unit_to_inch}
\alias{unit_to_inch}
\title{Convert a measurement to inches based on the unit}
\usage{
unit_to_inch(val, unit, dpi = 150)
}
\arguments{
\item{val}{The measurement value}

\item{unit}{The unit of the measurement (e.g., "px", "mm", "cm", "in")}

\item{dpi}{The resolution to use when interpreting pixel units}
}
\value{
The measurement converted to inches
}
\description{
Convert a measurement to inches based on the unit
}
