% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgedit-package.R
\docType{package}
\name{svgedit-package}
\alias{svgedit}
\alias{svgedit-package}
\title{svgedit: Insert Graphs, Images and Text in SVG Files}
\description{
Edit SVG files created in 'Inkscape' by replacing placeholders (e.g. a rectangle element or {} in a text box) by 'ggplot2' objects, images or text. This helps automate the creation of figures with complex layouts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DanielThedie/svgedit}
  \item Report bugs at \url{https://github.com/DanielThedie/svgedit/issues}
}

}
\author{
\strong{Maintainer}: Daniel Thedie \email{daniel.thedie@ed.ac.uk} (\href{https://orcid.org/0000-0002-1352-7245}{ORCID}) [copyright holder]

}
\keyword{internal}
